requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
debugpy==1.8.14
Django==5.2.4
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-filter==25.1
django-timezone-field==7.1
djangorestframework==3.16.0
docker==7.1.0
drf-yasg==1.21.10
gunicorn==23.0.0
h11==0.16.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
kombu==5.5.4
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2-binary==2.9.10
Pygments==2.19.2
pytest==8.4.1
pytest-celery==1.2.0
pytest_docker_tools==3.1.9
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
requests==2.32.4
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tenacity==9.1.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
